netease_vip_extractor/requirements.txt
requests==2.31.0
pycryptodome==3.19.0
prettytable==3.9.0 